fix(clippy): resolve all clippy warnings for CI compliance
Clippyエラー修正（GitHub CI対応）:

1. src/core/backup.rs:648
   - `match_result_ok`: `.ok()`冗長パターンマッチ修正
   - `Some(rel_path) = dest.strip_prefix().ok()` → `Ok(rel_path) = dest.strip_prefix()`

2. src/core/incremental.rs:40-44
   - `derivable_impls`: 手動Default実装を#[derive(Default)]に変更
   - #[default]属性でFull variantをデフォルト指定

3. src/core/incremental.rs:203
   - `unnecessary_map_or`: map_or簡略化
   - `map_or(true, |prev| prev != &current_hash)` → `!= Some(&current_hash)`

4. src/core/logging.rs:45,77
   - `should_implement_trait`: from_str → parse にリネーム
   - std::str::FromStrとの混同回避

5. benches/backup_benchmark.rs (6箇所)
   - let runner → let mut runner (E0596借用エラー修正)

6. benches/backup_performance.rs (3箇所)
   - let runner → let mut runner (E0596借用エラー修正)

7. tests/integrity_tests.rs:24
   - `needless_borrows_for_generic_args`: 不要な借用削除
   - &source_dir.join("subdir") → source_dir.join("subdir")

8. examples/test_progress.rs:26,64
   - `unnecessary_cast`: 不要なu64キャスト削除
   - ((x as u64) * ...) as u64 → (x as u64) * ...

修正理由:
- GitHub CI の cargo clippy -D warnings 完全対応
- Rust 1.91.0 Clippy新規lint対応

テスト: 全ファイルコンパイル確認済み

+++ examples/test_progress.rs
@@ -23,7 +23,7 @@ fn main() {
         // 統計情報の更新
         let elapsed = start.elapsed().as_secs_f64();
         let processed_files = i + 1;
-        let total_bytes = ((processed_files as u64) * 1024 * 1024) as u64; // 1MBずつ増加
+        let total_bytes = (processed_files as u64) * 1024 * 1024; // 1MBずつ増加
         progress.update_stats(processed_files as u64, total_bytes, elapsed);
 
         progress.inc(1);
@@ -61,7 +61,7 @@ fn main() {
         ));
 
         let elapsed = start.elapsed().as_secs_f64();
-        let total_bytes = ((i + 1) as u64 * 100 * 1024 * 1024) as u64; // 平均100MB
+        let total_bytes = (i + 1) as u64 * 100 * 1024 * 1024; // 平均100MB
         progress.update_stats((i + 1) as u64, total_bytes, elapsed);
 
         progress.inc(1);
